+++ test/Unit/FieldDefinition/OptionalTest.php
@@ -29,6 +29,7 @@ use Faker\Generator;
  * @uses \Ergebnis\FactoryBot\FieldDefinition
  * @uses \Ergebnis\FactoryBot\FieldDefinition\Value
  * @uses \Ergebnis\FactoryBot\FixtureFactory
+ * @uses \Ergebnis\FactoryBot\Strategy\DefaultStrategy
  */
 final class OptionalTest extends Unit\AbstractTestCase
 {