+++ test/Unit/FieldDefinition/ReferencesTest.php
@@ -30,6 +30,7 @@ use Example\Entity;
  * @uses \Ergebnis\FactoryBot\FieldDefinition
  * @uses \Ergebnis\FactoryBot\FieldDefinition\Value
  * @uses \Ergebnis\FactoryBot\FixtureFactory
+ * @uses \Ergebnis\FactoryBot\Strategy\DefaultStrategy
  */
 final class ReferencesTest extends Unit\AbstractTestCase
 {