+++ test/Unit/FixtureFactoryTest.php
@@ -26,6 +26,7 @@ use Faker\Generator;
  * @internal
  *
  * @covers \Ergebnis\FactoryBot\FixtureFactory
+ * @covers \Ergebnis\FactoryBot\Strategy\DefaultStrategy
  *
  * @uses \Ergebnis\FactoryBot\Count
  * @uses \Ergebnis\FactoryBot\EntityDefinition